Added test case and comments, renamed test file
Added test case and comments explaining test cases.

Changed name of test file to conform with convention.

Added numPoints as a parameter to LIC5.
